improvements part 1
