+++ mostlyai/sdk/_data/constraint_transformations.py
@@ -16,223 +16,28 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import logging
-from abc import ABC, abstractmethod
 from pathlib import Path
 
 import pandas as pd
 
+from mostlyai.sdk._data._constraint_types import (
+    ConstraintHandler,
+    FixedCombinationHandler,
+    InequalityHandler,
+)
 from mostlyai.sdk.client._constraint_types import (
     FixedCombination,
     Inequality,
     convert_constraint_config_to_typed,
 )
-from mostlyai.sdk.domain import (
-    Generator,
-    ModelEncodingType,
-)
+from mostlyai.sdk.domain import Generator
 
 _LOG = logging.getLogger(__name__)
 
 # type alias for constraint types
 ConstraintType = FixedCombination | Inequality
-
-
-def _generate_internal_column_name(prefix: str, columns: list[str]) -> str:
-    """generate a deterministic internal column name."""
-    key = "|".join(columns)
-    hash_suffix = hashlib.md5(key.encode()).hexdigest()[:8]
-    columns_str = "_".join(col.upper() for col in columns)
-    return f"__TABULAR_CONSTRAINT_{prefix}_{columns_str}_{hash_suffix}__"
-
-
-class ConstraintHandler(ABC):
-    """abstract base class for constraint handlers."""
-
-    @abstractmethod
-    def get_internal_column_names(self) -> list[str]:
-        """return list of internal column names created by this handler."""
-        pass
-
-    @abstractmethod
-    def to_internal(self, df: pd.DataFrame) -> pd.DataFrame:
-        """transform dataframe (in-place) from user schema to internal schema."""
-        pass
-
-    @abstractmethod
-    def to_original(self, df: pd.DataFrame) -> pd.DataFrame:
-        """transform dataframe (in-place) from internal schema back to user schema."""
-        pass
-
-    @abstractmethod
-    def get_encoding_types(self) -> dict[str, str]:
-        """return encoding types for internal columns."""
-        pass
-
-    def _validate_columns(self, df: pd.DataFrame, columns: list[str]) -> None:
-        """validate that all required columns exist in the dataframe."""
-        missing_cols = set(columns) - set(df.columns)
-        if missing_cols:
-            raise ValueError(
-                f"Columns {sorted(missing_cols)} required by {self.__class__.__name__} "
-                f"not found in dataframe. Available columns: {sorted(df.columns)}"
-            )
-
-
-class FixedCombinationHandler(ConstraintHandler):
-    """handler for FixedCombination constraints."""
-
-    def __init__(self, constraint: FixedCombination):
-        self.constraint = constraint
-        self.table_name = constraint.table_name
-        self.columns = constraint.columns
-        self.merged_name = _generate_internal_column_name("FC", self.columns)
-
-    def get_internal_column_names(self) -> list[str]:
-        return [self.merged_name]
-
-    def to_internal(self, df: pd.DataFrame) -> pd.DataFrame:
-        self._validate_columns(df, self.columns)
-
-        def merge_row(row):
-            values = [row[col] if pd.notna(row[col]) else None for col in self.columns]
-            # JSON serialization handles all escaping automatically
-            return json.dumps(values, ensure_ascii=False)
-
-        df[self.merged_name] = df.apply(merge_row, axis=1)
-        return df
-
-    def to_original(self, df: pd.DataFrame) -> pd.DataFrame:
-        if self.merged_name in df.columns:
-
-            def split_row(merged_value: str) -> list[str]:
-                if pd.isna(merged_value):
-                    return [""] * len(self.columns)
-                elif merged_value == "_RARE_":
-                    return ["_RARE_"] * len(self.columns)
-                try:
-                    values = json.loads(merged_value)
-                    return [str(v) if v is not None else "" for v in values]
-                except json.JSONDecodeError:
-                    _LOG.error(f"failed to decode JSON for {merged_value}; using empty values")
-                    return [""] * len(self.columns)
-
-            split_values = df[self.merged_name].apply(split_row)
-            split_df = pd.DataFrame(split_values.tolist(), index=df.index)
-
-            # preserve original index
-            original_index = df.index
-            split_df.index = original_index
-
-            # assign to original columns
-            for i, col in enumerate(self.columns):
-                df[col] = split_df[i].values
-
-            # drop the merged column
-            df = df.drop(columns=[self.merged_name])
-        return df
-
-    def get_encoding_types(self) -> dict[str, str]:
-        # always use TABULAR encoding for constraints, regardless of model_type
-        # constraints merge columns which requires categorical encoding
-        return {self.merged_name: "TABULAR_CATEGORICAL"}
-
-
-class InequalityHandler(ConstraintHandler):
-    """handler for Inequality constraints (low <= high)."""
-
-    _DATETIME_EPOCH = pd.Timestamp("1970-01-01")  # reference epoch for delta representation
-
-    def __init__(self, constraint: Inequality, table=None):
-        self.constraint = constraint
-        self.table_name = constraint.table_name
-        self.low_column = constraint.low_column
-        self.high_column = constraint.high_column
-        self._delta_column = _generate_internal_column_name("INEQ_DELTA", [self.low_column, self.high_column])
-
-        # determine if this is a datetime constraint based on table encoding types
-        self._is_datetime = False
-        if table and table.columns:
-            datetime_encodings = {
-                ModelEncodingType.tabular_datetime,
-                # ModelEncodingType.tabular_datetime_relative,  # not supported yet
-            }
-            # check if either column is datetime-encoded
-            self._is_datetime = all(
-                col.model_encoding_type in datetime_encodings
-                for col in table.columns
-                if col.name in {self.low_column, self.high_column}
-            )
-
-    def __repr__(self) -> str:
-        return f"{self.low_column} <= {self.high_column}"
-
-    def get_internal_column_names(self) -> list[str]:
-        return [self._delta_column]
-
-    def to_internal(self, df: pd.DataFrame) -> pd.DataFrame:
-        self._validate_columns(df, [self.low_column, self.high_column])
-        low = df[self.low_column]
-        high = df[self.high_column]
-
-        # handle NAs: if either low or high is NA, set delta to NA
-        na_mask = low.isna() | high.isna()
-
-        zero = pd.Timedelta(0) if self._is_datetime else 0
-        delta = high - low
-
-        # set delta to NA where either low or high is NA
-        delta[na_mask] = pd.NA
-
-        # only check violations where we have valid values
-        violations = (delta < zero) & ~na_mask
-        if violations.any():
-            _LOG.warning(f"detected {violations.sum() / len(delta) * 100:.2f}% inequality violations for {self}")
-
-        # convert timedelta to datetime using epoch (for datetime constraints)
-        if self._is_datetime:
-            # represent delta as datetime: epoch + timedelta
-            # preserve NAs during conversion
-            delta = delta.where(na_mask, self._DATETIME_EPOCH + delta)
-        # for numeric, keep as-is (numeric delta)
-
-        df[self._delta_column] = delta
-        return df
-
-    def to_original(self, df: pd.DataFrame) -> pd.DataFrame:
-        """transform from internal schema back to original schema."""
-        if self._delta_column not in df.columns:
-            return df
-
-        # prepare data and types
-        low = df[self.low_column]
-        delta = df[self._delta_column]
-        high_dtype = df[self.high_column].dtype
-
-        # convert datetime back to timedelta if needed
-        if self._is_datetime:
-            # ensure delta is datetime dtype before subtraction
-            delta = pd.to_datetime(delta)
-            # preserve NAs during conversion
-            delta = delta - self._DATETIME_EPOCH
-            delta = delta.astype("timedelta64[ns]")  # explicit dtype conversion
-
-        # reconstruction: high = low + delta, but if delta is NA, keep generated values
-        na_mask = delta.isna()
-        df[self.high_column] = df[self.high_column].where(na_mask, low + delta)
-        if pd.api.types.is_integer_dtype(high_dtype):
-            df[self.high_column] = df[self.high_column].astype(high_dtype)
-
-        return df.drop(columns=[self._delta_column])
-
-    def get_encoding_types(self) -> dict[str, str]:
-        # use TABULAR_DATETIME for datetime constraints to preserve precision
-        # use TABULAR_NUMERIC_AUTO for numeric constraints
-        if self._is_datetime:
-            return {self._delta_column: "TABULAR_DATETIME"}
-        return {self._delta_column: "TABULAR_NUMERIC_AUTO"}
 
 
 def _create_constraint_handler(constraint: ConstraintType, table=None) -> ConstraintHandler:
